Subject: Add curent header string to HeaderDialogFrag message
What Happened: Added a class variable in HeaderCursorAdapter to grab the header column string showing in the screen Pass thsi variable through to HeaderDialogFrag in the FragementManager tag. Now when doing an edit the column header string is shown in the HeaderDialogFrag view. Should have done this in a different branch than develop but forgot I was on develop.